% Generated by roxygen2 (4.0.0): do not edit by hand
\name{load_resultsets}
\alias{load_resultsets}
\title{Get the resultset table for a given commit or resultset_hash}
\usage{
load_resultsets(resultset_hash = NULL, commit = NULL)
}
\description{
Get the resultset table for a given commit or resultset_hash
}

