\name{nearest_commit_with_resultset}
\alias{nearest_commit_with_resultset}
\title{This function is only to be called from vdiffstat}
\usage{
  nearest_commit_with_resultset(start = "", n = 20)
}
\arguments{
  \item{start}{The commit to start searching backward from}

  \item{n}{Maximum number of commits to search}
}
\description{
  This function is only to be called from vdiffstat
}

