% Generated by roxygen2 (4.0.0): do not edit by hand
\name{recent_vtest}
\alias{recent_vtest}
\title{Returns a data frame of recent commits and their associated resultset_hash.}
\usage{
recent_vtest(dir = ".", start = "", n = 20, main_branch = TRUE)
}
\arguments{
  \item{start}{The commit to start searching backward from}

  \item{dir}{The directory with the git repository}

  \item{n}{Maximum number of commits to search}

  \item{main_branch}{Don't show commits on branches that
  were merged in.}
}
\description{
For commits without a resultset, the resultset_hash is \code{NA}.
}

