% Generated by roxygen2 (4.0.0): do not edit by hand
\name{vdiffstat}
\alias{vdiffstat}
\title{Find test results that changed between ref1 and ref2}
\usage{
vdiffstat(ref1 = "HEAD", ref2 = "", pkg = ".", filter = "",
  all = FALSE)
}
\arguments{
  \item{ref1}{a git commit ref to compare (usually this
  should be the older ref). If \code{"recent"}, then use
  the most recent commit with a saved resultset.}

  \item{ref2}{a git commit ref to compare (usually this
  should be the newer ref). The empty string \code{""}
  refers to the last-run tests.}

  \item{pkg}{package object or path.}

  \item{filter}{a regular expression; result pages will be
  generated only only for test contexts that match this
  pattern.}

  \item{all}{if \code{TRUE}, show only Added, Deleted, and
  Changed test items.  If \code{FALSE}, also show Unchanged
  test items.}
}
\value{
a data frame with information about test results that changed.
}
\description{
This function is used for comparing the test results for two different
commits of the package that was tested.
}
\seealso{
\code{\link{vdiff_webpage}}
}

