\docType{package}
\name{vtest}
\alias{package-vtest}
\alias{vtest}
\alias{vtest-package}
\title{vtest}
\usage{
  vtest(pkg = NULL, filter = "", showhelp = TRUE)
}
\arguments{
  \item{pkg}{a path to the package, or a package object}

  \item{filter}{only tests with file names matching this
  regular expression will be executed.}

  \item{showhelp}{if \code{TRUE}, print helpful messages.}
}
\description{
  vtest

  Run visual tests
}

